trocando int pra float

+++ first_list/exercise1.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 
-float get_expression_result(int x){
+float get_expression_result(float x){
 	//Returns always one, since both expression are equals, but a leave both just to ilustrate.
 	return (x *((-x) + (x * x) - (x * x * x) + (x * x * x * x))) / (-x *((x) - (x * x) + (x * x * x) - (x * x * x * x))) ;
 }
 
 int main(int argc, char const *argv[])
 {
-	int x;
-	scanf("%d",&x);
+	float x;
+	scanf("%f",&x);
 	printf("%f",get_expression_result(x));
 	return 0;
 }
